Add an Authentication middleware to protect the endpoints.
The idea here is that if no auth token is present in the gateway state,
then we can assume that the supervisor is running without authentication
and everything should continue to work as it always has. However, if we
do have an auth token in the state, then every HTTP request needs to
contain a fully formed Authorization header containing a Bearer token
that matches what we have in the state. Anything short of that results
in an Unauthorized response.

Signed-off-by: Josh Black <raskchanky@gmail.com>